requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
websockets==15.0
pynacl==1.5.0
